Use bigint for per-step collision counters to fix Natt overflow
The per-step collision counters (nattempt_one, ncollide_one, nreact_one)
were 32-bit ints, so at high density a single step could attempt more
than 2^31 collisions and Natt would roll over to a negative value
(reported in issue #492). The running totals were already bigint, but
they accumulate the already-overflowed per-step values.

Widen the per-step counters to bigint in the base Collide class and in
both Kokkos reduction paths:

- collide.h: ncollide_one/nattempt_one/nreact_one -> bigint
- COLLIDE_REDUCE (parallel_reduce value_type) fields -> bigint
- atomic path: move the 3 collision counters out of the int[11] scalar
  view into a separate bigint[3] view (d_counters), leaving the 8
  genuine int flags in d_scalars; add a tdual_bigint_scalar typedef.

Per-cell nattempt is left as int: a single cell exceeding 2^31 attempts
is computationally infeasible, and it feeds the loop bound / gpair int
arrays. Only the per-step sums overflow in practice.

Performance: the counters are touched once per cell and once per
collision inside the collision kernel, which dominates runtime (~91% in
a collision-heavy CPU/OpenMP benchmark). An A/B run of the parallel_reduce
path showed no measurable change (56.7s int vs 57.4s bigint Coll time,
within run-to-run noise).

+++ src/collide.h
@@ -31,7 +31,7 @@ class Collide : protected Pointers {
   int nearcp;         // 1 for near neighbor collisions
   int nearlimit;      // limit on neighbor serach for near neigh collisions
 
-  int ncollide_one,nattempt_one,nreact_one;
+  bigint ncollide_one,nattempt_one,nreact_one;
   bigint ncollide_running,nattempt_running,nreact_running;
 
   Collide(class SPARTA *, int, char **);
